Okay hear me out... Debugging the projectile emission code turned into a refactoring. So. This commit adds a test case for 2 trucks shooting a bullet once per second. The refactoring is about adding a name parameter to systems so we can more easily log them and I'm doubting the decision to integrate the ECS and the systems so tightly. Game now has pointers to all the system types. In the future we could consider moving the systems iterate and update loop out of the ecs. Why does it need to be general?

+++ src/core/types.h
@@ -13,7 +13,7 @@ typedef struct {
 } ImageMeta;
 
 typedef uint32_t AssetId;
-typedef uint8_t SignatureT;
+typedef uint32_t SignatureT;
 typedef int EntityId;
 typedef size_t EntityIndex;
